Move functions for parsing command line arguments to a separate file
